Query for pods in a 24 hour window and not in that instantaneous instant
note the [24h:5m] -- without that, you get instant data. This is why when you queried it, it gave you no results (it was in the future), then sometime later, Alex queries it, and that moment had elapsed, so it had data
2:48
I propose that you add an assertion to every prom query that makes sure you're adding the window
2:51
I can actually confirm the above theory about times based on your messages in slack. The timestamp 1752084000 is for 18:00:00 UTC July 9 --
@Manas Sivakumar
 your message about it not showing data was at 17:41:00 UTC July 9. Alex saying "I see data" happened at 18:55:00 UTC July 9.