Testing Querying of allocations per aggregate (#14)
* Add another field to AllocationRequest struct to include 'IncludeIdle'

* Check if the code works

* added a new test to bats

* corrected custom window test case if condition

* Added test cases for varying breakdowns and window values

added test description and assumptions as comments at the top of the file

* update comment on response data loop

* Added another tests for window size verification

* Added Another test to check if shareIdle works (part of it)

* commenting out test cases with longer thanger than a day window

validating new changes by alex

Signed-off-by: Manas Sivakumar <manas23601@gmail.com>

* update test cases

* Adding Network and LoadBalancer Cost

* test run got removed during merge

* added error margin to test

* Testing Querying of allocations per aggregate

* Allocations Summary API

* comparing /allocation and /allocation/summary data

this test is far from complete, I'm only checking the basic things in the description

* Trying to geground truth

* updated to run the right test file

* updated to run the right test file

* combined allocation api tests into one single directory

* print mismatch number

* did gofmt

* Compare costs between allocation and allocation summary

* Updated all the tests to print the results when they fail

* using git diff to visualize differences in podNames

* update test file description

* Tests for CPU and RAM Bytes

the tests are not complete yet, this is just skeleton code that fails outright

* changed prometheus test to use RAMByteHours

* tries using rambytehours on the last 24 hours

difference is huge

* make use promInput struct in cpu allocation test

* changes input type to promehtues

* promInput failure fixed]

* using ignore metrics option

* Test RAMByteHours code in a different testfile

* ram byte usage max

* add bats test for RAMMax Usage

* new tests

* Updated Consolidated_RAM_COSTS_ANALYSIS

* one single test for RAM Costs

figuring out the zero value costs

* Got values now, some of them are not matching

* Check for RAMByteRequestAverage

* renamed rambytehours to rambyteshours

* podMap should be based on pod

* got it to work

two bugs, 1. I used the filter unit=byte in allocation prom request. 2. failed to accound some namespaces have only one output

* Formatted files

* Added a difference tolerance of 0.00001 for all costs

* remove additional time condition in expected output

* RAM Max Usage Costs

* When test fails, prints difference values between expected and got

* I skip all pods with zero nil values

* Client Test flakiness resolved by sorting equalto and notequalto filter parts

* renamed files

* running test for RAM Max Usage per pod

* using GitDiff to print difference in results

* testing checkboxes in readme

* update promql query to check the last 24h window

* CPUCoreHours and CPUCoreRequestAverage Tests

* GPU Hours and GPU Request Average Tests

* remove duplicate test name

* [feat] CPU Avg Usage Test

Is it necessary for me to validate aggregates based on pod and namespace

* [feat] RAM Average Usage Test

* [feat] GPU Usage Average

* Run Average Cost Tests in pipeline

Edited test file name in assert/count/test.bats

* [feat] GPU Max Usage

* [feat] CPU Max Time Test

* Revert "[feat] CPU Max Time Test"

This reverts commit 57ce91e037920267f98c7282431d8ba42bad3dbe.

* [fix] resolve pod count by ignoring prometheus-system-unmounted-pvcs

* [fix] RAM Max Time needs to be done on container

* [Fix] GPU Max Time should be based on pod,contianer

* Update REAMDE with done tests

* [update feat] uncomment test cases with different window sizes

I uncommented them earlier because I was in a dilemma on whether to mark these tests valid or not. I guess we can have them anyway and add more later

* [fix] removes flakiness in pod count by mentioning end time

* [formatting] applied gofmt

* Query for pods in a 24 hour window and not in that instantaneous instant

note the [24h:5m] -- without that, you get instant data. This is why when you queried it, it gave you no results (it was in the future), then sometime later, Alex queries it, and that moment had elapsed, so it had data
2:48
I propose that you add an assertion to every prom query that makes sure you're adding the window
2:51
I can actually confirm the above theory about times based on your messages in slack. The timestamp 1752084000 is for 18:00:00 UTC July 9 --
@Manas Sivakumar
 your message about it not showing data was at 17:41:00 UTC July 9. Alex saying "I see data" happened at 18:55:00 UTC July 9.

* Query for pods in a 24 hour window and not in that instantaneous instant

note the [24h:5m] -- without that, you get instant data. This is why when you queried it, it gave you no results (it was in the future), then sometime later, Alex queries it, and that moment had elapsed, so it had data
2:48
I propose that you add an assertion to every prom query that makes sure you're adding the window
2:51
I can actually confirm the above theory about times based on your messages in slack. The timestamp 1752084000 is for 18:00:00 UTC July 9 --
@Manas Sivakumar
 your message about it not showing data was at 17:41:00 UTC July 9. Alex saying "I see data" happened at 18:55:00 UTC July 9.

* [modify] Rewrite Avg Usage Tests to account for PodRunTimes while calculating Averages

* [update] edit comment message to specify metric condition

* [format] do formatting

---------

Signed-off-by: Manas Sivakumar <manas23601@gmail.com>